requirements.txt
opencv-python-headless
opencv-contrib-python-headless
flask
flask_restful
werkzeug
pandas
paddlepaddle
paddleocr
pillow
pyperclip
splinter
pillow
